refactoring, created book and note classes
