+++ lib/rn/commands/notes.rb
@@ -13,30 +13,12 @@ module RN
           'thoughts --book Memoires    # Creates a note titled "thoughts" in the book "Memoires"'
         ]
 
-        require "tempfile"
+        require_relative "../classes/note"
 
         def call(title:, **options)
           book = options[:book]
-          if valid_name? title
-            title += ".rn"
-            if !book.nil?
-              if valid_name?(book)
-                Books::Create.new.call name: book
-              else
-                return puts "Only numbers, letters and spaces are allowed for the book title"
-              end
-            end
-            tmp = Tempfile.new("buffer")
-            tmp.rewind
-            TTY::Editor.open(tmp.path, command: default_editor)
-            if tmp.size > 0
-              TTY::File.copy_file tmp.path, files_path(book, title)
-            else
-              puts "You cannot create an empty note."
-            end
-          else
-            puts "Only numbers, letters and spaces are allowed for the note title"
-          end
+          title += ".rn"
+          Note.create(title, book)
         rescue Errno::EACCES
           puts "Permission denied for create '#{title}' on '#{books_path(book)}'"
         end
@@ -57,12 +39,7 @@ module RN
         def call(title:, **options)
           book = options[:book]
           title += ".rn"
-          if File.exist? files_path(book, title)
-            TTY::File.remove_file files_path(book, title)
-            puts "File '#{title}' removed succesfully."
-          else
-            puts "Error: File '#{title}' does not exist."
-          end
+          Note.delete(title, book)
         end
       end
 
@@ -81,12 +58,7 @@ module RN
         def call(title:, **options)
           book = options[:book]
           title += ".rn"
-          if File.exist? files_path(book, title)
-            TTY::Editor.open(files_path(book, title), command: default_editor)
-            puts "File '#{title}' updated succesfully."
-          else
-            puts "Error: File #{title} does not exist."
-          end
+          Note.edit(title, book)
         end
       end
 
@@ -107,16 +79,7 @@ module RN
           book = options[:book]
           old_title += ".rn"
           new_title += ".rn"
-          if File.exist? files_path(book, old_title)
-            if File.exist? files_path(book, new_title)
-              puts "Error: File #{new_title} already exists, please try with another name"
-            else
-              File.rename(files_path(book, old_title), files_path(book, new_title))
-              puts "Renamed file #{old_title} to #{new_title}"
-            end
-          else
-            puts "Error: File #{old_title} does not exist."
-          end
+          Note.retitle(old_title, new_title, book)
         end
       end
 
@@ -169,7 +132,7 @@ module RN
           book = options[:book]
           title += ".rn"
           if File.exist? files_path(book, title)
-            File.foreach(files_path(book, title)) { |line| puts line }
+            Note.show(title, book)
           else
             puts "Error: file '#{title}' does not exist"
           end
